requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
audioop-lts==0.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
Django==6.0
django-jazzmin==3.0.1
fonttools==4.61.1
freetype-py==2.5.1
gTTS==2.5.4
html5lib==1.1
idna==3.11
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
narwhals==2.15.0
numpy==2.4.1
oscrypto==1.3.0
packaging==25.0
pillow==12.1.0
plotly==6.5.2
pycairo==1.29.0
pycparser==3.0
pydub==0.25.1
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyMuPDF==1.26.7
pyparsing==3.3.1
pypdf==6.6.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
PyYAML==6.0.3
reportlab==4.4.9
requests==2.32.5
rlPyCairo==0.4.0
setuptools==80.9.0
six==1.17.0
SpeechRecognition==3.14.5
sqlparse==0.5.5
standard-aifc==3.13.0
standard-chunk==3.13.0
svglib==1.6.0
tinycss2==1.5.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.2
webencodings==0.5.1
wheel==0.45.1
xhtml2pdf==0.2.17
